requirement.txt
pyTelegramBotAPI
load_dotenv
asyncio
aiohttp
pep505
asyncssh